+++ oak-core/src/main/java/org/apache/jackrabbit/oak/query/ast/CoalesceImpl.java
@@ -140,13 +140,4 @@ public class CoalesceImpl extends DynamicOperandImpl {
         return null;
     }
 
-    @Override
-    public String getOrderEntryPropertyName(SelectorImpl s) {
-        String fn = getFunction(s);
-        if (fn != null) {
-            return  QueryConstants.FUNCTION_RESTRICTION_PREFIX + fn;
-        }
-        return null;
-    }
-
 }
