+++ oak-core/src/main/java/org/apache/jackrabbit/oak/query/ast/LengthImpl.java
@@ -156,13 +156,4 @@ public class LengthImpl extends DynamicOperandImpl {
         return null;
     }
 
-    @Override
-    public String getOrderEntryPropertyName(SelectorImpl s) {
-        String fn = getFunction(s);
-        if (fn != null) {
-            return  QueryConstants.FUNCTION_RESTRICTION_PREFIX + fn;
-        }
-        return null;
-    }
-
 }
