+++ oak-core/src/main/java/org/apache/jackrabbit/oak/query/ast/NodeLocalNameImpl.java
@@ -149,13 +149,4 @@ public class NodeLocalNameImpl extends DynamicOperandImpl {
             OrderEntry.Order.DESCENDING : OrderEntry.Order.ASCENDING);
     }
 
-    @Override
-    public String getOrderEntryPropertyName(SelectorImpl s) {
-        String fn = getFunction(s);
-        if (fn != null) {
-            return  QueryConstants.FUNCTION_RESTRICTION_PREFIX + fn;
-        }
-        return null;
-    }
-
 }
