+++ oak-core/src/main/java/org/apache/jackrabbit/oak/query/ast/NodeNameImpl.java
@@ -192,13 +192,4 @@ public class NodeNameImpl extends DynamicOperandImpl {
             OrderEntry.Order.DESCENDING : OrderEntry.Order.ASCENDING);
     }
 
-    @Override
-    public String getOrderEntryPropertyName(SelectorImpl s) {
-        String fn = getFunction(s);
-        if (fn != null) {
-            return  QueryConstants.FUNCTION_RESTRICTION_PREFIX + fn;
-        }
-        return null;
-    }
-
 }
