+++ oak-core/src/main/java/org/apache/jackrabbit/oak/query/ast/UpperCaseImpl.java
@@ -137,13 +137,4 @@ public class UpperCaseImpl extends DynamicOperandImpl {
         return null;
     }
 
-    @Override
-    public String getOrderEntryPropertyName(SelectorImpl s) {
-        String fn = getFunction(s);
-        if (fn != null) {
-            return  QueryConstants.FUNCTION_RESTRICTION_PREFIX + fn;
-        }
-        return null;
-    }
-
 }
